pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "herocheck"
version = "0.3.0"
description = "B2B SaaS hero section analyzer — persona-driven buying committee simulation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "firecrawl-py>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
herocheck = "herocheck.cli:main"

[tool.setuptools.packages.find]
include = ["herocheck*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
